+++ src/weapons/index.ts
@@ -1,0 +1,2 @@
+export { BaseWeapon } from './BaseWeapon';
+export { Pistol } from './Pistol';